improvement: caching
During testing a bug I was not able to set a breakpoint because the
scripts were loaded using another URL during reload. So I changed the
caching functionality to generate a new URL every 10000 seconds. This
should avoid errors due to caching on the users site without bothering
them with long loading times